refactor: import svgs from deriv x
