+++ src/images/svg/deriv-x/index.ts
@@ -1,0 +1,25 @@
+import Icon24_7 from 'images/svg/deriv-x/24-7.svg'
+import IconMultipleMarketing from 'images/svg/deriv-x/multiple-marketing.svg'
+import IconNewPromising from 'images/svg/deriv-x/new-promising.svg'
+import mt5Icon from 'images/svg/deriv-x/mt5.svg'
+import dtIcon from 'images/svg/deriv-x/dt-icosn.svg'
+import ezIcon from 'images/svg/deriv-x/ez-icon.svg'
+import dGoIcon from 'images/svg/deriv-x/go-icon.svg'
+import dbIcon from 'images/svg/deriv-x/db-icon.svg'
+import derivXLogo from 'images/svg/deriv-x/derivX.svg'
+import derivXQR from 'images/svg/deriv-x/derivxQR.png'
+import DerivXLogo from 'images/svg/deriv-x/deriv-x-banner-logo.svg'
+
+export {
+    Icon24_7,
+    IconMultipleMarketing,
+    IconNewPromising,
+    mt5Icon,
+    dtIcon,
+    ezIcon,
+    dGoIcon,
+    dbIcon,
+    derivXLogo,
+    derivXQR,
+    DerivXLogo,
+}
