+++ src/pages/derivx/_get-derivx.tsx
@@ -1,13 +1,12 @@
 import React from 'react'
 import styled from 'styled-components'
-import derivXLogo from '../../images/svg/deriv-x/derivX.svg'
-import derivXQR from '../../images/svg/deriv-x/derivxQR.png'
 import GrayAngle30 from '../../images/common/gray-angle.png'
 import AppleIcon from '../../images/svg/apple-icon.svg'
 import AppGalleryIcon from '../../images/svg/appGallery-icon.svg'
 import AndroidIcon from '../../images/svg/android-icon.svg'
 import BrowserIcon from '../../images/svg/browser-icon.svg'
 import GetAppMobileBG from '../../images/common/getAppMobileBG.png'
+import { derivXLogo, derivXQR } from 'images/svg/deriv-x'
 import CommonHeaderSection from 'components/elements/common-header-section'
 import MultiWidthColumn from 'components/elements/multi-width-column'
 import device from 'themes/device'
