+++ src/pages/derivx/_hero.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { StaticImage } from 'gatsby-plugin-image'
 import CommonHeaderSection from 'components/elements/common-header-section'
-import DerivXLogo from 'images/svg/deriv-x/deriv-x-banner-logo.svg'
+import { DerivXLogo } from 'images/svg/deriv-x'
 import device from 'themes/device'
 import useBreakpoints from 'components/hooks/use-breakpoints'
 import useHandleSignup from 'components/hooks/use-handle-signup'
